LC-4sum [Runtime: N/A; Memory: N/A]

+++ sorting/4sum.cpp
@@ -9,7 +9,7 @@ public:
             for(int j=i+1;j<n;j++) {
                 int l=j+1;
                 int r=n-1;
-                long int sum=target-nums[i]-nums[j];
+                long long sum=target-nums[i]-nums[j];
                 while(l<r) {
                     int total=nums[l]+nums[r];
                     if(total==sum) {
